requirements.txt
aiohttp==3.13.3
py-markdown-table==1.3.0
aiofiles==25.1.0
PyGithub==2.8.1